examples/create-table-02/requirements.txt
cassandra-driver==3.16.0
numpy==1.15.4
pandas==0.23.4
python-dateutil==2.7.5
pytz==2018.7
six==1.11.0
tqdm==4.28.1
